Revert "🚚💥 Refactor: move files to src, eslint current files, add config exemple (#21)"
This reverts commit 63e5c95fb7ddb7224ac231d46dbcf90f1724d762.
