--- src/index.js
@@ -1,33 +1,0 @@
-require('dotenv').config();
-const { Client, Collection } = require('discord.js');
-const fs = require('fs');
-const client = new Client({
-  allowedMentions: {
-    parse: ['roles', 'users'],
-  },
-  partials: ['MESSAGE', 'CHANNEL', 'REACTION'],
-  intents: [
-    'GUILDS',
-    'GUILD_MEMBERS',
-    'GUILD_BANS',
-    'GUILD_EMOJIS_AND_STICKERS',
-    'GUILD_VOICE_STATES',
-    'GUILD_MESSAGES',
-    'GUILD_MESSAGE_REACTIONS',
-    'DIRECT_MESSAGES',
-  ],
-});
-module.exports = client;
-
-client.commands = new Collection();
-client.aliases = new Collection();
-client.events = new Collection();
-client.cooldowns = new Collection();
-client.slashCommands = new Collection();
-client.categories = fs.readdirSync('./commands/');
-
-['command_handler', 'event_handler', 'slash_handler'].forEach((handler) => {
-  require(`./handlers/${handler}`)(client);
-});
-
-client.login(process.env.TOKEN);
